all graphql schemas created and some fixes

+++ src/main/java/com/vitavault/vitavault/domain/SubscriptionPlan.java
@@ -16,7 +16,7 @@ public class SubscriptionPlan extends BaseEntity {
     private String name;
 
     @Column(nullable = false)
-    private Double cost;
+    private float cost;
 
     @Column(nullable = false)
     private String terms;
